Organize identityHashCode/synchronized deprecations
