+++ compose/runtime/runtime/src/jvmMain/kotlin/androidx/compose/runtime/OldSynchronization.jvm.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 The Android Open Source Project
+ * Copyright 2024 The Android Open Source Project
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,10 +14,18 @@
  * limitations under the License.
  */
 
+@file:JvmName("ActualJvm_jvmKt")
+@file:JvmMultifileClass
+
 package androidx.compose.runtime
 
-// TODO https://youtrack.jetbrains.com/issue/CMP-719/Make-expect-fun-identityHashCodeinstance-Any-Int-internal
-@InternalComposeApi
-@Deprecated("Made internal. It wasn't supposed to be public")
-fun identityHashCode(instance: Any?): Int  =
-    androidx.compose.runtime.internal.identityHashCode(instance)
+import androidx.compose.runtime.platform.SynchronizedObject
+
+@PublishedApi
+@JvmName("synchronized")
+@Deprecated(
+    level = DeprecationLevel.HIDDEN,
+    message = "not expected to be referenced directly as the old version had to be inlined"
+)
+internal inline fun <R> oldSynchronized(lock: SynchronizedObject, block: () -> R): R =
+    androidx.compose.runtime.platform.synchronized(lock, block)
