+++ compose/runtime/runtime/src/wasmJsMain/kotlin/androidx/compose/runtime/OldIdentityHashCode.wasmJs.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 The Android Open Source Project
+ * Copyright 2025 The Android Open Source Project
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,8 +16,10 @@
 
 package androidx.compose.runtime
 
-// TODO https://youtrack.jetbrains.com/issue/CMP-719/Make-expect-fun-identityHashCodeinstance-Any-Int-internal
 @InternalComposeApi
-@Deprecated("Made internal. It wasn't supposed to be public")
+@Deprecated(
+    level = DeprecationLevel.HIDDEN,
+    message = "Made internal. It wasn't supposed to be public"
+)
 fun identityHashCode(instance: Any?): Int  =
     androidx.compose.runtime.internal.identityHashCode(instance)
